Add option to not enable SELinux booleans
Previously it always set these two variables if SELinux was enabled, allowing
Samba to share any file in the system (so the "setype" share option is not
really used because Samba can access any type file).

I would prefer to disable these by default, but maybe it's better not to break
existing users?